refactor de logica para poder guardar la cantidad de los insumos con dos decimales

+++ src/database/migrations/20240514195754-create-insumos.js
@@ -17,7 +17,7 @@ module.exports = {
         type: Sequelize.STRING
       },
       cantidad: {
-        type: Sequelize.INTEGER
+        type: Sequelize.DECIMAL(10,2)
       },
       idProducto:{
         type:Sequelize.INTEGER,
